+++ src/routes/store.js
@@ -5,3 +5,4 @@ export const updateRate = writable(15000);
 export const entryStore = writable(10);
 export const sortDirStore = writable("ascending");
 export const sortByStore = writable("rank");
+export const dataLoading = writable(true);